feat: qa testing enhancements
