+++ js/FeedzyLoop/edit.js
@@ -2,8 +2,6 @@
 /**
  * WordPress dependencies.
  */
-import { __ } from '@wordpress/i18n';
-
 import {
 	store as blocksStore,
 	createBlocksFromInnerBlocksTemplate,
@@ -14,18 +12,8 @@ import {
 	store as blockEditorStore,
 	__experimentalBlockVariationPicker as BlockVariationPicker,
 	useBlockProps,
-	BlockControls,
 	InnerBlocks,
-	InspectorControls,
 } from '@wordpress/block-editor';
-
-import {
-	PanelBody,
-	RangeControl,
-	SelectControl,
-	ToolbarButton,
-	ToolbarGroup,
-} from '@wordpress/components';
 
 import { useDispatch, useSelect } from '@wordpress/data';
 
@@ -38,7 +26,7 @@ import ServerSideRender from '@wordpress/server-side-render';
  */
 import metadata from './block.json';
 import Placeholder from './placeholder';
-import ConditionsControl from '../Conditions/ConditionsControl';
+import Controls from './controls';
 
 const { name } = metadata;
 
@@ -109,154 +97,6 @@ const Edit = ({ attributes, setAttributes, clientId }) => {
 		});
 	};
 
-	const Controls = () => (
-		<>
-			<BlockControls>
-				<ToolbarGroup>
-					<ToolbarButton
-						icon="edit"
-						title={__('Edit Feed', 'feedzy-rss-feeds')}
-						onClick={() => setIsEditing(true)}
-					/>
-				</ToolbarGroup>
-
-				<ToolbarGroup>
-					<ToolbarButton
-						onClick={() => setIsPreviewing(!isPreviewing)}
-					>
-						{isPreviewing
-							? __('Hide Preview', 'feedzy-rss-feeds')
-							: __('Show Preview', 'feedzy-rss-feeds')}
-					</ToolbarButton>
-				</ToolbarGroup>
-			</BlockControls>
-
-			<InspectorControls>
-				<PanelBody title={__('Settings', 'feedzy-rss-feeds')}>
-					<RangeControl
-						label={__('Column Count', 'feedzy-rss-feeds')}
-						value={attributes?.layout?.columnCount || 1}
-						onChange={(value) =>
-							onChangeLayout({ type: 'columnCount', value })
-						}
-						min={1}
-						max={5}
-					/>
-
-					<RangeControl
-						label={__('Number of Items', 'feedzy-rss-feeds')}
-						value={attributes?.query?.max || 5}
-						onChange={(value) =>
-							onChangeQuery({ type: 'max', value })
-						}
-						min={1}
-						max={20}
-					/>
-
-					<SelectControl
-						label={__('Sorting Order', 'feedzy-rss-feeds')}
-						value={attributes?.query?.sort}
-						options={[
-							{
-								label: __('Default', 'feedzy-rss-feeds'),
-								value: 'default',
-							},
-							{
-								label: __(
-									'Date Descending',
-									'feedzy-rss-feeds'
-								),
-								value: 'date_desc',
-							},
-							{
-								label: __('Date Ascending', 'feedzy-rss-feeds'),
-								value: 'date_asc',
-							},
-							{
-								label: __(
-									'Title Descending',
-									'feedzy-rss-feeds'
-								),
-								value: 'title_desc',
-							},
-							{
-								label: __(
-									'Title Ascending',
-									'feedzy-rss-feeds'
-								),
-								value: 'title_asc',
-							},
-						]}
-						onChange={(value) =>
-							onChangeQuery({ type: 'sort', value })
-						}
-					/>
-
-					<SelectControl
-						label={__('Feed Caching Time', 'feedzy-rss-feeds')}
-						value={attributes?.query?.refresh || '12_hours'}
-						options={[
-							{
-								label: __('1 Hour', 'feedzy-rss-feeds'),
-								value: '1_hours',
-							},
-							{
-								label: __('2 Hours', 'feedzy-rss-feeds'),
-								value: '3_hours',
-							},
-							{
-								label: __('12 Hours', 'feedzy-rss-feeds'),
-								value: '12_hours',
-							},
-							{
-								label: __('1 Day', 'feedzy-rss-feeds'),
-								value: '1_days',
-							},
-							{
-								label: __('3 Days', 'feedzy-rss-feeds'),
-								value: '3_days',
-							},
-							{
-								label: __('15 Days', 'feedzy-rss-feeds'),
-								value: '15_days',
-							},
-						]}
-						onChange={(value) =>
-							onChangeQuery({ type: 'refresh', value })
-						}
-					/>
-				</PanelBody>
-
-				<PanelBody
-					title={[
-						__('Filter items', 'feedzy-rss-feeds'),
-						!feedzyData.isPro && (
-							<span className="fz-pro-label">Pro</span>
-						),
-					]}
-					initialOpen={false}
-					className={
-						feedzyData.isPro
-							? 'feedzy-item-filter'
-							: 'feedzy-item-filter fz-locked'
-					}
-				>
-					<ConditionsControl
-						conditions={
-							attributes?.conditions || {
-								conditions: [],
-								match: 'all',
-							}
-						}
-						setConditions={(conditions) => {
-							setAttributes({ conditions });
-						}}
-					/>
-				</PanelBody>
-			</InspectorControls>
-		</>
-	);
-
 	if (isEditing) {
 		return (
 			<div {...blockProps}>
@@ -272,7 +112,16 @@ const Edit = ({ attributes, setAttributes, clientId }) => {
 	if ((!isSelected || isPreviewing) && innerBlocksContent) {
 		return (
 			<>
-				<Controls />
+				<Controls
+					attributes={attributes}
+					isEditing={isEditing}
+					isPreviewing={isPreviewing}
+					setAttributes={setAttributes}
+					onChangeLayout={onChangeLayout}
+					onChangeQuery={onChangeQuery}
+					setIsEditing={setIsEditing}
+					setIsPreviewing={setIsPreviewing}
+				/>
 
 				<div {...blockProps}>
 					<ServerSideRender
@@ -289,7 +138,16 @@ const Edit = ({ attributes, setAttributes, clientId }) => {
 
 	return (
 		<>
-			<Controls />
+			<Controls
+				attributes={attributes}
+				isEditing={isEditing}
+				isPreviewing={isPreviewing}
+				setAttributes={setAttributes}
+				onChangeLayout={onChangeLayout}
+				onChangeQuery={onChangeQuery}
+				setIsEditing={setIsEditing}
+				setIsPreviewing={setIsPreviewing}
+			/>
 
 			<div {...blockProps}>
 				{hasInnerBlocks ? (
