+++ js/FeedzyLoop/editor.scss
@@ -3,6 +3,11 @@
         .components-base-control {
             width: 100%;
         }
+
+        p {
+            margin: 0;
+            width: 100%;
+        }
     }
 }
